@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.31.0
-// 	protoc        v4.24.4
+// 	protoc-gen-go v1.36.6
+// 	protoc        v6.30.2
 // source: github.com/dogmatiq/protean/internal/proteanpb/error.proto
 
 package proteanpb
@@ -12,6 +12,7 @@ import (
 	anypb "google.golang.org/protobuf/types/known/anypb"
 	reflect "reflect"
 	sync "sync"
+	unsafe "unsafe"
 )
 
 const (
@@ -27,22 +28,19 @@ const (
 // When serialized using the protojson package it conforms to the structure of a
 // JSON-RPC v2 error object. See https://www.jsonrpc.org/specification#error_object.
 type Error struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Code          int32                  `protobuf:"varint,1,opt,name=code,proto3" json:"code,omitempty"`
+	Message       string                 `protobuf:"bytes,2,opt,name=message,proto3" json:"message,omitempty"`
+	Data          *anypb.Any             `protobuf:"bytes,3,opt,name=data,proto3" json:"data,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Code    int32      `protobuf:"varint,1,opt,name=code,proto3" json:"code,omitempty"`
-	Message string     `protobuf:"bytes,2,opt,name=message,proto3" json:"message,omitempty"`
-	Data    *anypb.Any `protobuf:"bytes,3,opt,name=data,proto3" json:"data,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *Error) Reset() {
 	*x = Error{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_github_com_dogmatiq_protean_internal_proteanpb_error_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_github_com_dogmatiq_protean_internal_proteanpb_error_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Error) String() string {
@@ -53,7 +51,7 @@ func (*Error) ProtoMessage() {}
 
 func (x *Error) ProtoReflect() protoreflect.Message {
 	mi := &file_github_com_dogmatiq_protean_internal_proteanpb_error_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -92,20 +90,17 @@ func (x *Error) GetData() *anypb.Any {
 // SupportedMediaTypes is a list of media types supported by the server, in
 // order of preference.
 type SupportedMediaTypes struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	MediaTypes    []string               `protobuf:"bytes,1,rep,name=media_types,json=mediaTypes,proto3" json:"media_types,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	MediaTypes []string `protobuf:"bytes,1,rep,name=media_types,json=mediaTypes,proto3" json:"media_types,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *SupportedMediaTypes) Reset() {
 	*x = SupportedMediaTypes{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_github_com_dogmatiq_protean_internal_proteanpb_error_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_github_com_dogmatiq_protean_internal_proteanpb_error_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *SupportedMediaTypes) String() string {
@@ -116,7 +111,7 @@ func (*SupportedMediaTypes) ProtoMessage() {}
 
 func (x *SupportedMediaTypes) ProtoReflect() protoreflect.Message {
 	mi := &file_github_com_dogmatiq_protean_internal_proteanpb_error_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -140,43 +135,32 @@ func (x *SupportedMediaTypes) GetMediaTypes() []string {
 
 var File_github_com_dogmatiq_protean_internal_proteanpb_error_proto protoreflect.FileDescriptor
 
-var file_github_com_dogmatiq_protean_internal_proteanpb_error_proto_rawDesc = []byte{
-	0x0a, 0x3a, 0x67, 0x69, 0x74, 0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d, 0x2f, 0x64, 0x6f, 0x67,
-	0x6d, 0x61, 0x74, 0x69, 0x71, 0x2f, 0x70, 0x72, 0x6f, 0x74, 0x65, 0x61, 0x6e, 0x2f, 0x69, 0x6e,
-	0x74, 0x65, 0x72, 0x6e, 0x61, 0x6c, 0x2f, 0x70, 0x72, 0x6f, 0x74, 0x65, 0x61, 0x6e, 0x70, 0x62,
-	0x2f, 0x65, 0x72, 0x72, 0x6f, 0x72, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12, 0x0a, 0x70, 0x72,
-	0x6f, 0x74, 0x65, 0x61, 0x6e, 0x2e, 0x76, 0x31, 0x1a, 0x19, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65,
-	0x2f, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2f, 0x61, 0x6e, 0x79, 0x2e, 0x70, 0x72,
-	0x6f, 0x74, 0x6f, 0x22, 0x5f, 0x0a, 0x05, 0x45, 0x72, 0x72, 0x6f, 0x72, 0x12, 0x12, 0x0a, 0x04,
-	0x63, 0x6f, 0x64, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x05, 0x52, 0x04, 0x63, 0x6f, 0x64, 0x65,
-	0x12, 0x18, 0x0a, 0x07, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28,
-	0x09, 0x52, 0x07, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x12, 0x28, 0x0a, 0x04, 0x64, 0x61,
-	0x74, 0x61, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x14, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c,
-	0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x41, 0x6e, 0x79, 0x52, 0x04,
-	0x64, 0x61, 0x74, 0x61, 0x22, 0x36, 0x0a, 0x13, 0x53, 0x75, 0x70, 0x70, 0x6f, 0x72, 0x74, 0x65,
-	0x64, 0x4d, 0x65, 0x64, 0x69, 0x61, 0x54, 0x79, 0x70, 0x65, 0x73, 0x12, 0x1f, 0x0a, 0x0b, 0x6d,
-	0x65, 0x64, 0x69, 0x61, 0x5f, 0x74, 0x79, 0x70, 0x65, 0x73, 0x18, 0x01, 0x20, 0x03, 0x28, 0x09,
-	0x52, 0x0a, 0x6d, 0x65, 0x64, 0x69, 0x61, 0x54, 0x79, 0x70, 0x65, 0x73, 0x42, 0x30, 0x5a, 0x2e,
-	0x67, 0x69, 0x74, 0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d, 0x2f, 0x64, 0x6f, 0x67, 0x6d, 0x61,
-	0x74, 0x69, 0x71, 0x2f, 0x70, 0x72, 0x6f, 0x74, 0x65, 0x61, 0x6e, 0x2f, 0x69, 0x6e, 0x74, 0x65,
-	0x72, 0x6e, 0x61, 0x6c, 0x2f, 0x70, 0x72, 0x6f, 0x74, 0x65, 0x61, 0x6e, 0x70, 0x62, 0x62, 0x06,
-	0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
-}
+const file_github_com_dogmatiq_protean_internal_proteanpb_error_proto_rawDesc = "" +
+	"\n" +
+	":github.com/dogmatiq/protean/internal/proteanpb/error.proto\x12\n" +
+	"protean.v1\x1a\x19google/protobuf/any.proto\"_\n" +
+	"\x05Error\x12\x12\n" +
+	"\x04code\x18\x01 \x01(\x05R\x04code\x12\x18\n" +
+	"\amessage\x18\x02 \x01(\tR\amessage\x12(\n" +
+	"\x04data\x18\x03 \x01(\v2\x14.google.protobuf.AnyR\x04data\"6\n" +
+	"\x13SupportedMediaTypes\x12\x1f\n" +
+	"\vmedia_types\x18\x01 \x03(\tR\n" +
+	"mediaTypesB0Z.github.com/dogmatiq/protean/internal/proteanpbb\x06proto3"
 
 var (
 	file_github_com_dogmatiq_protean_internal_proteanpb_error_proto_rawDescOnce sync.Once
-	file_github_com_dogmatiq_protean_internal_proteanpb_error_proto_rawDescData = file_github_com_dogmatiq_protean_internal_proteanpb_error_proto_rawDesc
+	file_github_com_dogmatiq_protean_internal_proteanpb_error_proto_rawDescData []byte
 )
 
 func file_github_com_dogmatiq_protean_internal_proteanpb_error_proto_rawDescGZIP() []byte {
 	file_github_com_dogmatiq_protean_internal_proteanpb_error_proto_rawDescOnce.Do(func() {
-		file_github_com_dogmatiq_protean_internal_proteanpb_error_proto_rawDescData = protoimpl.X.CompressGZIP(file_github_com_dogmatiq_protean_internal_proteanpb_error_proto_rawDescData)
+		file_github_com_dogmatiq_protean_internal_proteanpb_error_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_github_com_dogmatiq_protean_internal_proteanpb_error_proto_rawDesc), len(file_github_com_dogmatiq_protean_internal_proteanpb_error_proto_rawDesc)))
 	})
 	return file_github_com_dogmatiq_protean_internal_proteanpb_error_proto_rawDescData
 }
 
 var file_github_com_dogmatiq_protean_internal_proteanpb_error_proto_msgTypes = make([]protoimpl.MessageInfo, 2)
-var file_github_com_dogmatiq_protean_internal_proteanpb_error_proto_goTypes = []interface{}{
+var file_github_com_dogmatiq_protean_internal_proteanpb_error_proto_goTypes = []any{
 	(*Error)(nil),               // 0: protean.v1.Error
 	(*SupportedMediaTypes)(nil), // 1: protean.v1.SupportedMediaTypes
 	(*anypb.Any)(nil),           // 2: google.protobuf.Any
@@ -195,37 +179,11 @@ func file_github_com_dogmatiq_protean_internal_proteanpb_error_proto_init() {
 	if File_github_com_dogmatiq_protean_internal_proteanpb_error_proto != nil {
 		return
 	}
-	if !protoimpl.UnsafeEnabled {
-		file_github_com_dogmatiq_protean_internal_proteanpb_error_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Error); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_github_com_dogmatiq_protean_internal_proteanpb_error_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*SupportedMediaTypes); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
-			RawDescriptor: file_github_com_dogmatiq_protean_internal_proteanpb_error_proto_rawDesc,
+			RawDescriptor: unsafe.Slice(unsafe.StringData(file_github_com_dogmatiq_protean_internal_proteanpb_error_proto_rawDesc), len(file_github_com_dogmatiq_protean_internal_proteanpb_error_proto_rawDesc)),
 			NumEnums:      0,
 			NumMessages:   2,
 			NumExtensions: 0,
@@ -236,7 +194,6 @@ func file_github_com_dogmatiq_protean_internal_proteanpb_error_proto_init() {
 		MessageInfos:      file_github_com_dogmatiq_protean_internal_proteanpb_error_proto_msgTypes,
 	}.Build()
 	File_github_com_dogmatiq_protean_internal_proteanpb_error_proto = out.File
-	file_github_com_dogmatiq_protean_internal_proteanpb_error_proto_rawDesc = nil
 	file_github_com_dogmatiq_protean_internal_proteanpb_error_proto_goTypes = nil
 	file_github_com_dogmatiq_protean_internal_proteanpb_error_proto_depIdxs = nil
 }
