@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.31.0
-// 	protoc        v4.24.4
+// 	protoc-gen-go v1.36.6
+// 	protoc        v6.30.2
 // source: github.com/dogmatiq/protean/internal/stringservice/service.proto
 
 package stringservice
@@ -11,6 +11,7 @@ import (
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 	reflect "reflect"
 	sync "sync"
+	unsafe "unsafe"
 )
 
 const (
@@ -23,20 +24,17 @@ const (
 // ToUpperRequest is the RPC input message for the StringService.ToUpper() RPC
 // method.
 type ToUpperRequest struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
-	OriginalString string `protobuf:"bytes,1,opt,name=original_string,json=originalString,proto3" json:"original_string,omitempty"`
+	state          protoimpl.MessageState `protogen:"open.v1"`
+	OriginalString string                 `protobuf:"bytes,1,opt,name=original_string,json=originalString,proto3" json:"original_string,omitempty"`
+	unknownFields  protoimpl.UnknownFields
+	sizeCache      protoimpl.SizeCache
 }
 
 func (x *ToUpperRequest) Reset() {
 	*x = ToUpperRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_github_com_dogmatiq_protean_internal_stringservice_service_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_github_com_dogmatiq_protean_internal_stringservice_service_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *ToUpperRequest) String() string {
@@ -47,7 +45,7 @@ func (*ToUpperRequest) ProtoMessage() {}
 
 func (x *ToUpperRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_github_com_dogmatiq_protean_internal_stringservice_service_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -72,20 +70,17 @@ func (x *ToUpperRequest) GetOriginalString() string {
 // ToUpperResponse is the RPC output message for the StringService.ToUpper() RPC
 // method.
 type ToUpperResponse struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
-	UppercaseString string `protobuf:"bytes,1,opt,name=uppercase_string,json=uppercaseString,proto3" json:"uppercase_string,omitempty"`
+	state           protoimpl.MessageState `protogen:"open.v1"`
+	UppercaseString string                 `protobuf:"bytes,1,opt,name=uppercase_string,json=uppercaseString,proto3" json:"uppercase_string,omitempty"`
+	unknownFields   protoimpl.UnknownFields
+	sizeCache       protoimpl.SizeCache
 }
 
 func (x *ToUpperResponse) Reset() {
 	*x = ToUpperResponse{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_github_com_dogmatiq_protean_internal_stringservice_service_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_github_com_dogmatiq_protean_internal_stringservice_service_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *ToUpperResponse) String() string {
@@ -96,7 +91,7 @@ func (*ToUpperResponse) ProtoMessage() {}
 
 func (x *ToUpperResponse) ProtoReflect() protoreflect.Message {
 	mi := &file_github_com_dogmatiq_protean_internal_stringservice_service_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -120,46 +115,30 @@ func (x *ToUpperResponse) GetUppercaseString() string {
 
 var File_github_com_dogmatiq_protean_internal_stringservice_service_proto protoreflect.FileDescriptor
 
-var file_github_com_dogmatiq_protean_internal_stringservice_service_proto_rawDesc = []byte{
-	0x0a, 0x40, 0x67, 0x69, 0x74, 0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d, 0x2f, 0x64, 0x6f, 0x67,
-	0x6d, 0x61, 0x74, 0x69, 0x71, 0x2f, 0x70, 0x72, 0x6f, 0x74, 0x65, 0x61, 0x6e, 0x2f, 0x69, 0x6e,
-	0x74, 0x65, 0x72, 0x6e, 0x61, 0x6c, 0x2f, 0x73, 0x74, 0x72, 0x69, 0x6e, 0x67, 0x73, 0x65, 0x72,
-	0x76, 0x69, 0x63, 0x65, 0x2f, 0x73, 0x65, 0x72, 0x76, 0x69, 0x63, 0x65, 0x2e, 0x70, 0x72, 0x6f,
-	0x74, 0x6f, 0x12, 0x0e, 0x70, 0x72, 0x6f, 0x74, 0x65, 0x61, 0x6e, 0x2e, 0x73, 0x74, 0x72, 0x69,
-	0x6e, 0x67, 0x22, 0x39, 0x0a, 0x0e, 0x54, 0x6f, 0x55, 0x70, 0x70, 0x65, 0x72, 0x52, 0x65, 0x71,
-	0x75, 0x65, 0x73, 0x74, 0x12, 0x27, 0x0a, 0x0f, 0x6f, 0x72, 0x69, 0x67, 0x69, 0x6e, 0x61, 0x6c,
-	0x5f, 0x73, 0x74, 0x72, 0x69, 0x6e, 0x67, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x0e, 0x6f,
-	0x72, 0x69, 0x67, 0x69, 0x6e, 0x61, 0x6c, 0x53, 0x74, 0x72, 0x69, 0x6e, 0x67, 0x22, 0x3c, 0x0a,
-	0x0f, 0x54, 0x6f, 0x55, 0x70, 0x70, 0x65, 0x72, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65,
-	0x12, 0x29, 0x0a, 0x10, 0x75, 0x70, 0x70, 0x65, 0x72, 0x63, 0x61, 0x73, 0x65, 0x5f, 0x73, 0x74,
-	0x72, 0x69, 0x6e, 0x67, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x0f, 0x75, 0x70, 0x70, 0x65,
-	0x72, 0x63, 0x61, 0x73, 0x65, 0x53, 0x74, 0x72, 0x69, 0x6e, 0x67, 0x32, 0x5b, 0x0a, 0x0d, 0x53,
-	0x74, 0x72, 0x69, 0x6e, 0x67, 0x53, 0x65, 0x72, 0x76, 0x69, 0x63, 0x65, 0x12, 0x4a, 0x0a, 0x07,
-	0x54, 0x6f, 0x55, 0x70, 0x70, 0x65, 0x72, 0x12, 0x1e, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x65, 0x61,
-	0x6e, 0x2e, 0x73, 0x74, 0x72, 0x69, 0x6e, 0x67, 0x2e, 0x54, 0x6f, 0x55, 0x70, 0x70, 0x65, 0x72,
-	0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x1f, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x65, 0x61,
-	0x6e, 0x2e, 0x73, 0x74, 0x72, 0x69, 0x6e, 0x67, 0x2e, 0x54, 0x6f, 0x55, 0x70, 0x70, 0x65, 0x72,
-	0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x42, 0x34, 0x5a, 0x32, 0x67, 0x69, 0x74, 0x68,
-	0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d, 0x2f, 0x64, 0x6f, 0x67, 0x6d, 0x61, 0x74, 0x69, 0x71, 0x2f,
-	0x70, 0x72, 0x6f, 0x74, 0x65, 0x61, 0x6e, 0x2f, 0x69, 0x6e, 0x74, 0x65, 0x72, 0x6e, 0x61, 0x6c,
-	0x2f, 0x73, 0x74, 0x72, 0x69, 0x6e, 0x67, 0x73, 0x65, 0x72, 0x76, 0x69, 0x63, 0x65, 0x62, 0x06,
-	0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
-}
+const file_github_com_dogmatiq_protean_internal_stringservice_service_proto_rawDesc = "" +
+	"\n" +
+	"@github.com/dogmatiq/protean/internal/stringservice/service.proto\x12\x0eprotean.string\"9\n" +
+	"\x0eToUpperRequest\x12'\n" +
+	"\x0foriginal_string\x18\x01 \x01(\tR\x0eoriginalString\"<\n" +
+	"\x0fToUpperResponse\x12)\n" +
+	"\x10uppercase_string\x18\x01 \x01(\tR\x0fuppercaseString2[\n" +
+	"\rStringService\x12J\n" +
+	"\aToUpper\x12\x1e.protean.string.ToUpperRequest\x1a\x1f.protean.string.ToUpperResponseB4Z2github.com/dogmatiq/protean/internal/stringserviceb\x06proto3"
 
 var (
 	file_github_com_dogmatiq_protean_internal_stringservice_service_proto_rawDescOnce sync.Once
-	file_github_com_dogmatiq_protean_internal_stringservice_service_proto_rawDescData = file_github_com_dogmatiq_protean_internal_stringservice_service_proto_rawDesc
+	file_github_com_dogmatiq_protean_internal_stringservice_service_proto_rawDescData []byte
 )
 
 func file_github_com_dogmatiq_protean_internal_stringservice_service_proto_rawDescGZIP() []byte {
 	file_github_com_dogmatiq_protean_internal_stringservice_service_proto_rawDescOnce.Do(func() {
-		file_github_com_dogmatiq_protean_internal_stringservice_service_proto_rawDescData = protoimpl.X.CompressGZIP(file_github_com_dogmatiq_protean_internal_stringservice_service_proto_rawDescData)
+		file_github_com_dogmatiq_protean_internal_stringservice_service_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_github_com_dogmatiq_protean_internal_stringservice_service_proto_rawDesc), len(file_github_com_dogmatiq_protean_internal_stringservice_service_proto_rawDesc)))
 	})
 	return file_github_com_dogmatiq_protean_internal_stringservice_service_proto_rawDescData
 }
 
 var file_github_com_dogmatiq_protean_internal_stringservice_service_proto_msgTypes = make([]protoimpl.MessageInfo, 2)
-var file_github_com_dogmatiq_protean_internal_stringservice_service_proto_goTypes = []interface{}{
+var file_github_com_dogmatiq_protean_internal_stringservice_service_proto_goTypes = []any{
 	(*ToUpperRequest)(nil),  // 0: protean.string.ToUpperRequest
 	(*ToUpperResponse)(nil), // 1: protean.string.ToUpperResponse
 }
@@ -178,37 +157,11 @@ func file_github_com_dogmatiq_protean_internal_stringservice_service_proto_init(
 	if File_github_com_dogmatiq_protean_internal_stringservice_service_proto != nil {
 		return
 	}
-	if !protoimpl.UnsafeEnabled {
-		file_github_com_dogmatiq_protean_internal_stringservice_service_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*ToUpperRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_github_com_dogmatiq_protean_internal_stringservice_service_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*ToUpperResponse); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
-			RawDescriptor: file_github_com_dogmatiq_protean_internal_stringservice_service_proto_rawDesc,
+			RawDescriptor: unsafe.Slice(unsafe.StringData(file_github_com_dogmatiq_protean_internal_stringservice_service_proto_rawDesc), len(file_github_com_dogmatiq_protean_internal_stringservice_service_proto_rawDesc)),
 			NumEnums:      0,
 			NumMessages:   2,
 			NumExtensions: 0,
@@ -219,7 +172,6 @@ func file_github_com_dogmatiq_protean_internal_stringservice_service_proto_init(
 		MessageInfos:      file_github_com_dogmatiq_protean_internal_stringservice_service_proto_msgTypes,
 	}.Build()
 	File_github_com_dogmatiq_protean_internal_stringservice_service_proto = out.File
-	file_github_com_dogmatiq_protean_internal_stringservice_service_proto_rawDesc = nil
 	file_github_com_dogmatiq_protean_internal_stringservice_service_proto_goTypes = nil
 	file_github_com_dogmatiq_protean_internal_stringservice_service_proto_depIdxs = nil
 }
