@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.31.0
-// 	protoc        v4.24.4
+// 	protoc-gen-go v1.36.6
+// 	protoc        v6.30.2
 // source: github.com/dogmatiq/protean/internal/testservice/service.proto
 
 package testservice
@@ -11,6 +11,7 @@ import (
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 	reflect "reflect"
 	sync "sync"
+	unsafe "unsafe"
 )
 
 const (
@@ -23,21 +24,18 @@ const (
 // Input is the message used as inputs to all of the RPC methods in the test
 // service.
 type Input struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Id            string                 `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`
+	Data          string                 `protobuf:"bytes,2,opt,name=data,proto3" json:"data,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Id   string `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`
-	Data string `protobuf:"bytes,2,opt,name=data,proto3" json:"data,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *Input) Reset() {
 	*x = Input{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_github_com_dogmatiq_protean_internal_testservice_service_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_github_com_dogmatiq_protean_internal_testservice_service_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Input) String() string {
@@ -48,7 +46,7 @@ func (*Input) ProtoMessage() {}
 
 func (x *Input) ProtoReflect() protoreflect.Message {
 	mi := &file_github_com_dogmatiq_protean_internal_testservice_service_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -80,21 +78,18 @@ func (x *Input) GetData() string {
 // Output is the message used as outputs from all of the RPC methods in the test
 // service.
 type Output struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Id            string                 `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`
+	Data          string                 `protobuf:"bytes,2,opt,name=data,proto3" json:"data,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Id   string `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`
-	Data string `protobuf:"bytes,2,opt,name=data,proto3" json:"data,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *Output) Reset() {
 	*x = Output{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_github_com_dogmatiq_protean_internal_testservice_service_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_github_com_dogmatiq_protean_internal_testservice_service_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Output) String() string {
@@ -105,7 +100,7 @@ func (*Output) ProtoMessage() {}
 
 func (x *Output) ProtoReflect() protoreflect.Message {
 	mi := &file_github_com_dogmatiq_protean_internal_testservice_service_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -136,54 +131,35 @@ func (x *Output) GetData() string {
 
 var File_github_com_dogmatiq_protean_internal_testservice_service_proto protoreflect.FileDescriptor
 
-var file_github_com_dogmatiq_protean_internal_testservice_service_proto_rawDesc = []byte{
-	0x0a, 0x3e, 0x67, 0x69, 0x74, 0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d, 0x2f, 0x64, 0x6f, 0x67,
-	0x6d, 0x61, 0x74, 0x69, 0x71, 0x2f, 0x70, 0x72, 0x6f, 0x74, 0x65, 0x61, 0x6e, 0x2f, 0x69, 0x6e,
-	0x74, 0x65, 0x72, 0x6e, 0x61, 0x6c, 0x2f, 0x74, 0x65, 0x73, 0x74, 0x73, 0x65, 0x72, 0x76, 0x69,
-	0x63, 0x65, 0x2f, 0x73, 0x65, 0x72, 0x76, 0x69, 0x63, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f,
-	0x12, 0x0c, 0x70, 0x72, 0x6f, 0x74, 0x65, 0x61, 0x6e, 0x2e, 0x74, 0x65, 0x73, 0x74, 0x22, 0x2b,
-	0x0a, 0x05, 0x49, 0x6e, 0x70, 0x75, 0x74, 0x12, 0x0e, 0x0a, 0x02, 0x69, 0x64, 0x18, 0x01, 0x20,
-	0x01, 0x28, 0x09, 0x52, 0x02, 0x69, 0x64, 0x12, 0x12, 0x0a, 0x04, 0x64, 0x61, 0x74, 0x61, 0x18,
-	0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x04, 0x64, 0x61, 0x74, 0x61, 0x22, 0x2c, 0x0a, 0x06, 0x4f,
-	0x75, 0x74, 0x70, 0x75, 0x74, 0x12, 0x0e, 0x0a, 0x02, 0x69, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28,
-	0x09, 0x52, 0x02, 0x69, 0x64, 0x12, 0x12, 0x0a, 0x04, 0x64, 0x61, 0x74, 0x61, 0x18, 0x02, 0x20,
-	0x01, 0x28, 0x09, 0x52, 0x04, 0x64, 0x61, 0x74, 0x61, 0x32, 0x81, 0x02, 0x0a, 0x0b, 0x54, 0x65,
-	0x73, 0x74, 0x53, 0x65, 0x72, 0x76, 0x69, 0x63, 0x65, 0x12, 0x32, 0x0a, 0x05, 0x55, 0x6e, 0x61,
-	0x72, 0x79, 0x12, 0x13, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x65, 0x61, 0x6e, 0x2e, 0x74, 0x65, 0x73,
-	0x74, 0x2e, 0x49, 0x6e, 0x70, 0x75, 0x74, 0x1a, 0x14, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x65, 0x61,
-	0x6e, 0x2e, 0x74, 0x65, 0x73, 0x74, 0x2e, 0x4f, 0x75, 0x74, 0x70, 0x75, 0x74, 0x12, 0x3b, 0x0a,
-	0x0c, 0x43, 0x6c, 0x69, 0x65, 0x6e, 0x74, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x12, 0x13, 0x2e,
-	0x70, 0x72, 0x6f, 0x74, 0x65, 0x61, 0x6e, 0x2e, 0x74, 0x65, 0x73, 0x74, 0x2e, 0x49, 0x6e, 0x70,
-	0x75, 0x74, 0x1a, 0x14, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x65, 0x61, 0x6e, 0x2e, 0x74, 0x65, 0x73,
-	0x74, 0x2e, 0x4f, 0x75, 0x74, 0x70, 0x75, 0x74, 0x28, 0x01, 0x12, 0x3b, 0x0a, 0x0c, 0x53, 0x65,
-	0x72, 0x76, 0x65, 0x72, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x12, 0x13, 0x2e, 0x70, 0x72, 0x6f,
-	0x74, 0x65, 0x61, 0x6e, 0x2e, 0x74, 0x65, 0x73, 0x74, 0x2e, 0x49, 0x6e, 0x70, 0x75, 0x74, 0x1a,
-	0x14, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x65, 0x61, 0x6e, 0x2e, 0x74, 0x65, 0x73, 0x74, 0x2e, 0x4f,
-	0x75, 0x74, 0x70, 0x75, 0x74, 0x30, 0x01, 0x12, 0x44, 0x0a, 0x13, 0x42, 0x69, 0x64, 0x69, 0x72,
-	0x65, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x61, 0x6c, 0x53, 0x74, 0x72, 0x65, 0x61, 0x6d, 0x12, 0x13,
-	0x2e, 0x70, 0x72, 0x6f, 0x74, 0x65, 0x61, 0x6e, 0x2e, 0x74, 0x65, 0x73, 0x74, 0x2e, 0x49, 0x6e,
-	0x70, 0x75, 0x74, 0x1a, 0x14, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x65, 0x61, 0x6e, 0x2e, 0x74, 0x65,
-	0x73, 0x74, 0x2e, 0x4f, 0x75, 0x74, 0x70, 0x75, 0x74, 0x28, 0x01, 0x30, 0x01, 0x42, 0x32, 0x5a,
-	0x30, 0x67, 0x69, 0x74, 0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d, 0x2f, 0x64, 0x6f, 0x67, 0x6d,
-	0x61, 0x74, 0x69, 0x71, 0x2f, 0x70, 0x72, 0x6f, 0x74, 0x65, 0x61, 0x6e, 0x2f, 0x69, 0x6e, 0x74,
-	0x65, 0x72, 0x6e, 0x61, 0x6c, 0x2f, 0x74, 0x65, 0x73, 0x74, 0x73, 0x65, 0x72, 0x76, 0x69, 0x63,
-	0x65, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
-}
+const file_github_com_dogmatiq_protean_internal_testservice_service_proto_rawDesc = "" +
+	"\n" +
+	">github.com/dogmatiq/protean/internal/testservice/service.proto\x12\fprotean.test\"+\n" +
+	"\x05Input\x12\x0e\n" +
+	"\x02id\x18\x01 \x01(\tR\x02id\x12\x12\n" +
+	"\x04data\x18\x02 \x01(\tR\x04data\",\n" +
+	"\x06Output\x12\x0e\n" +
+	"\x02id\x18\x01 \x01(\tR\x02id\x12\x12\n" +
+	"\x04data\x18\x02 \x01(\tR\x04data2\x81\x02\n" +
+	"\vTestService\x122\n" +
+	"\x05Unary\x12\x13.protean.test.Input\x1a\x14.protean.test.Output\x12;\n" +
+	"\fClientStream\x12\x13.protean.test.Input\x1a\x14.protean.test.Output(\x01\x12;\n" +
+	"\fServerStream\x12\x13.protean.test.Input\x1a\x14.protean.test.Output0\x01\x12D\n" +
+	"\x13BidirectionalStream\x12\x13.protean.test.Input\x1a\x14.protean.test.Output(\x010\x01B2Z0github.com/dogmatiq/protean/internal/testserviceb\x06proto3"
 
 var (
 	file_github_com_dogmatiq_protean_internal_testservice_service_proto_rawDescOnce sync.Once
-	file_github_com_dogmatiq_protean_internal_testservice_service_proto_rawDescData = file_github_com_dogmatiq_protean_internal_testservice_service_proto_rawDesc
+	file_github_com_dogmatiq_protean_internal_testservice_service_proto_rawDescData []byte
 )
 
 func file_github_com_dogmatiq_protean_internal_testservice_service_proto_rawDescGZIP() []byte {
 	file_github_com_dogmatiq_protean_internal_testservice_service_proto_rawDescOnce.Do(func() {
-		file_github_com_dogmatiq_protean_internal_testservice_service_proto_rawDescData = protoimpl.X.CompressGZIP(file_github_com_dogmatiq_protean_internal_testservice_service_proto_rawDescData)
+		file_github_com_dogmatiq_protean_internal_testservice_service_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_github_com_dogmatiq_protean_internal_testservice_service_proto_rawDesc), len(file_github_com_dogmatiq_protean_internal_testservice_service_proto_rawDesc)))
 	})
 	return file_github_com_dogmatiq_protean_internal_testservice_service_proto_rawDescData
 }
 
 var file_github_com_dogmatiq_protean_internal_testservice_service_proto_msgTypes = make([]protoimpl.MessageInfo, 2)
-var file_github_com_dogmatiq_protean_internal_testservice_service_proto_goTypes = []interface{}{
+var file_github_com_dogmatiq_protean_internal_testservice_service_proto_goTypes = []any{
 	(*Input)(nil),  // 0: protean.test.Input
 	(*Output)(nil), // 1: protean.test.Output
 }
@@ -208,37 +184,11 @@ func file_github_com_dogmatiq_protean_internal_testservice_service_proto_init() 
 	if File_github_com_dogmatiq_protean_internal_testservice_service_proto != nil {
 		return
 	}
-	if !protoimpl.UnsafeEnabled {
-		file_github_com_dogmatiq_protean_internal_testservice_service_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Input); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_github_com_dogmatiq_protean_internal_testservice_service_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Output); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
-			RawDescriptor: file_github_com_dogmatiq_protean_internal_testservice_service_proto_rawDesc,
+			RawDescriptor: unsafe.Slice(unsafe.StringData(file_github_com_dogmatiq_protean_internal_testservice_service_proto_rawDesc), len(file_github_com_dogmatiq_protean_internal_testservice_service_proto_rawDesc)),
 			NumEnums:      0,
 			NumMessages:   2,
 			NumExtensions: 0,
@@ -249,7 +199,6 @@ func file_github_com_dogmatiq_protean_internal_testservice_service_proto_init() 
 		MessageInfos:      file_github_com_dogmatiq_protean_internal_testservice_service_proto_msgTypes,
 	}.Build()
 	File_github_com_dogmatiq_protean_internal_testservice_service_proto = out.File
-	file_github_com_dogmatiq_protean_internal_testservice_service_proto_rawDesc = nil
 	file_github_com_dogmatiq_protean_internal_testservice_service_proto_goTypes = nil
 	file_github_com_dogmatiq_protean_internal_testservice_service_proto_depIdxs = nil
 }
